# Perf fix for encode and decode to use bytes data type 