As part of Antrea benchmarking, we found that bytes.Buffer type is incurring very high overhead in terms of allocations and memory consumed. Moving to bytes data type instead of `bytes.Buffer` wrapper.

Fixes: #205 

Before fix:
```
vagrantÉk8s-node-worker-1:ü/antrea/pkg/flowaggregator$ go test -test.v -run=none -test.benchmem  -bench=. -memprofile memprofile.out -cpuprofile profile.out
goos: linux
goarch: amd64
pkg: antrea.io/antrea/pkg/flowaggregator
cpu: Intel(R) Core(TM) i9-9980HK CPU É 2.40GHz
BenchmarkIntraNodeFlowRecords
    flowaggregator_perf_test.go:446: Num messages received: 1516755
BenchmarkIntraNodeFlowRecords-2   	       1	120001712897 ns/op	22902440680 B/op	431854675 allocs/op
PASS
ok  	antrea.io/antrea/pkg/flowaggregator	120.119s
```

After this fix:
```
vagrant@k8s-node-worker-1:~/antrea/pkg/flowaggregator$ go test -test.v -run=none -test.benchmem  -bench=. -memprofile memprofile.out -cpuprofile profile.out
goos: linux
goarch: amd64
pkg: antrea.io/antrea/pkg/flowaggregator
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkIntraNodeFlowRecords
    flowaggregator_perf_test.go:446: Num messages received: 1526508
BenchmarkIntraNodeFlowRecords-2   	       1	120005595596 ns/op	19760001232 B/op	356061206 allocs/op
PASS
ok  	antrea.io/antrea/pkg/flowaggregator	120.328s
```