@srikartati 2021-06-01
> Overall LGTM. Thanks for the changes. Encoding, decoding function and the tests looks much cleaner.

Thanks for the review, Yiou.
I think we can optimize further by doing pre-allocation of bytes slice in the record by using total elements size. This will improve the performance further. Will take it up in future PR.